+++ ui/package.json
@@ -3,12 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@chainsafe/filecoindot-types": "^0.1.12",
     "@emotion/react": "^11.7.0",
     "@emotion/styled": "^11.6.0",
     "@mui/icons-material": "^5.2.0",
     "@mui/material": "^5.2.2",
     "@mui/styled-engine-sc": "^5.1.0",
-    "@chainsafe/filecoindot-types": "^0.1.12",
     "@polkadot/api": "^7.2.1",
     "@polkadot/extension-dapp": "^0.42.4",
     "@polkadot/react-identicon": "^0.88.1",
@@ -28,12 +28,12 @@
     "web-vitals": "^1.0.1"
   },
   "devDependencies": {
-    "eslint": "^6.8.0",
+    "@babel/plugin-proposal-private-methods": "^7.16.7",
+    "@typescript-eslint/eslint-plugin": "^4.15.2",
+    "@typescript-eslint/parser": "^4.15.2",
     "eslint-plugin-react": "^7.27.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-ternary": "^1.0.4",
-    "@typescript-eslint/eslint-plugin": "^4.15.2",
-    "@typescript-eslint/parser": "^4.15.2"
+    "eslint-plugin-ternary": "^1.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
